Post feed adjustments - BottomSheetDialog and fake Database test

+++ app/src/main/java/co/pvitor/instagram/common/view/ModalBottomSheetDialog.kt
@@ -44,7 +44,7 @@ class ModalBottomSheetDialog: BottomSheetDialogFragment() {
             )
 
             textView.apply {
-                id = bottomSheetItem.textViewRes
+                id = bottomSheetItem.textViewRes.toInt()
                 text = context?.getText(bottomSheetItem.textViewRes)
                 setOnClickListener(itemListener)
             }
